Move `MapProof` building logic to separate module
The new approach to proof building decouples proofs from `ProofMapIndex`
and thus allows publishing proofs as a separate crate (e.g., for
WASM-powered light client).
