+++ components/merkledb/src/proof_map_index/mod.rs
@@ -30,7 +30,7 @@ use exonum_crypto::Hash;
 
 use self::{
     key::{BitsRange, ChildKind, VALUE_KEY_PREFIX},
-    proof::{create_multiproof, create_proof},
+    proof_builder::{BuildProof, MerklePatricia},
 };
 use crate::views::{AnyObject, IndexAddress};
 use crate::{
@@ -43,8 +43,29 @@ use crate::{
 mod key;
 mod node;
 mod proof;
+mod proof_builder;
 #[cfg(test)]
 mod tests;
+
+// Necessary to allow building proofs.
+impl<T, K, V> MerklePatricia<K, V> for ProofMapIndex<T, K, V>
+where
+    T: IndexAccess,
+    K: BinaryKey + ObjectHash,
+    V: BinaryValue,
+{
+    fn root_node(&self) -> Option<(ProofPath, Node)> {
+        self.get_root_node()
+    }
+
+    fn node(&self, path: &ProofPath) -> Node {
+        self.get_node_unchecked(path)
+    }
+
+    fn value(&self, key: &K) -> V {
+        self.get_value_unchecked(key)
+    }
+}
 
 /// A Merkelized version of a map that provides proofs of existence or non-existence for the map
 /// keys.
@@ -394,12 +415,7 @@ where
     /// let proof = index.get_proof(Hash::default());
     /// ```
     pub fn get_proof(&self, key: K) -> MapProof<K, V> {
-        create_proof(
-            key,
-            self.get_root_node(),
-            |path| self.get_node_unchecked(path),
-            |key| self.get_value_unchecked(key),
-        )
+        self.create_proof(key)
     }
 
     /// Returns the combined proof of existence or non-existence for the multiple specified keys.
@@ -419,12 +435,7 @@ where
     where
         KI: IntoIterator<Item = K>,
     {
-        create_multiproof(
-            keys,
-            self.get_root_node(),
-            |path| self.get_node_unchecked(path),
-            |key| self.get_value_unchecked(key),
-        )
+        self.create_multi_proof(keys)
     }
 
     /// Returns an iterator over the entries of the map in ascending order. The iterator element
